experiment(proof,caching): Add proof verified cache

+++ partner-chains/toolkit/data-sources/mock/src/block.rs
@@ -50,7 +50,9 @@ impl BlockDataSourceMock {
 	) -> Result<Option<MainchainBlock>> {
 		// reverse of computation in `get_latest_stable_block_for`
 		let block_number = u32::from_be_bytes(hash.0[..4].try_into().unwrap());
-		let timestamp = block_number * 20000;
+		// Widen before multiplying: at real-world timestamps `block_number` is ~millions, and
+		// `block_number * 20000` overflows u32 (panics under debug overflow checks).
+		let timestamp = block_number as u64 * 20000;
 		let epoch = block_number / self.block_per_epoch();
 		Ok(Some(MainchainBlock {
 			number: McBlockNumber(block_number),
